requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
astunparse==1.6.3
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cryptography==41.0.7
cssselect2==0.7.0
distro==1.9.0
Django==5.0.1
Flask==3.0.1
Flask-Markdown==0.3
flatbuffers==23.5.26
fonttools==4.47.2
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.14.0
libclang==16.0.6
Markdown==3.5
markdown-it-py==3.0.0
markdown-pdf==1.0
markdown2==2.4.12
Markdown2PDF==0.1.4
MarkupSafe==2.1.4
mdurl==0.1.2
ml-dtypes==0.2.0
numpy==1.26.1
oauthlib==3.2.2
openai==1.2.0
opt-einsum==3.3.0
packaging==23.2
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.3
pillow==10.2.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydyf==0.8.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyPDF2==3.0.1
pypdfium2==4.26.0
pyphen==0.14.0
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metal==1.1.0
termcolor==2.3.0
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
weasyprint==60.2
webencodings==0.5.1
Werkzeug==3.0.0
wrapt==1.14.1
zopfli==0.2.3
